Use pip3 not pip in the Dockerfile
When using 'python:3.7-slim-buster' the pip is a symlink to pip3,
and python is symlink to python3 (all in /usr/local/bin).

But if using base image like 'debian:buster-slim' or bullseye-slim and using python3 and python3-pip from Debian,
pip and python always symlink to python2 (and that is a good thing). And if python-pip
or python2 are not installed possibly in Debian bullseye), python and pip will not even exist.

So this does not change anything for the official build, or ones based on used python base image,
but allows to also use Debian images directly.